+++ src/main/java/com/walhay/gregifiedenergistics/mixins/gtceu/metatileentities/DataBankMixin.java
@@ -1,4 +1,4 @@
-package com.walhay.gregifiedenergistics.mixins.gtceu;
+package com.walhay.gregifiedenergistics.mixins.gtceu.metatileentities;
 
 import com.walhay.gregifiedenergistics.api.capability.IOpticalDataHandler;
 import com.walhay.gregifiedenergistics.mixins.interfaces.IDataBankUpdateHandler;
