+++ src/main/resources/mixins.ge.gtceu.json
@@ -5,14 +5,14 @@
   "minVersion" : "0.8",
   "compatibilityLevel" : "JAVA_8",
   "mixins" : [
-    "AssemblyLineMixin",
-	"OpticalNetHandlerMixin",
-	"DataAccessHatchMixin",
-	"OpticalDataHatchMixin",
-	"DataBankMixin",
-	"RecipeMixin",
-	"RecipeMapMixin",
-	"BlockOpticalPipeMixin",
+    "metatileentities.AssemblyLineMixin",
+	"metatileentities.DataAccessHatchMixin",
+	"metatileentities.DataBankMixin",
+	"pipenet.OpticalDataHatchMixin",
+	"pipenet.BlockOpticalPipeMixin",
+	"pipenet.OpticalNetHandlerMixin",
+	"recipes.RecipeMixin",
+	"recipes.RecipeMapMixin",
   ],
   "client": [],
   "server" : []
